moved VolunteerCategories to utilities for less confusion
